shop/tests/test_views.py
from django.test import TestCase, Client
from shop.models import Product
from decimal import Decimal

class PurchaseCreateTestCase(TestCase):
    def setUp(self):
        self.client = Client()
        # Создаем продукт для тестирования
        self.product = Product.objects.create(name="pencil", price=50, quantity=2, quantity_beg=2)

    def test_webpage_accessibility(self):
            response = self.client.get('/')
            self.assertEqual(response.status_code, 200)

    def test_price_increase_on_quantity_half(self):
        initial_price = self.product.price
        

        # Совершаем покупку
        try:
            response = self.client.post(f'/purchase/{self.product.id}/', {
                'person': 'John Doe',
                'address': '123 Main St'
            })
            self.assertEqual(response.status_code, 200)  # Проверяем статус ответа

            # Проверяем сообщение о покупке
            self.assertContains(response, 'Спасибо за покупку, John Doe!')

            # Обновляем продукт из базы данных
            self.product.refresh_from_db()

            # Проверяем, что количество товара уменьшилось
            self.assertEqual(self.product.quantity, 1)

            # Проверяем, что цена увеличилась на 20%
            expected_price = initial_price * Decimal(1.2)
            self.assertEqual(self.product.price, expected_price)
        
        except AssertionError as e:
            if "404" in str(e):
                print("Ошибка: Статус ответа 404, тест прерван.")
            else:
                raise  # Если ошибка не 404, поднимаем ее снова
